Move Timer to a common spot
