+++ examples/render_pyramid.py
@@ -20,6 +20,7 @@ from mercantile import Tile
 
 from marblecutter import tiling
 from marblecutter.sources import MemoryAdapter
+from marblecutter.stats import Timer
 from marblecutter.formats import PNG, GeoTIFF
 from marblecutter.transformations import Normal, Terrarium
 
@@ -45,16 +46,6 @@ RENDER_COMBINATIONS = [
     ("terrarium", TERRARIUM_TRANSFORMATION, PNG_FORMAT, ".png"),
     ("geotiff", None, GEOTIFF_FORMAT, ".tif"),
 ]
-
-
-class Timer(object):
-    def __enter__(self):
-        self.start = time.time()
-        return self
-
-    def __exit__(self, ty, val, tb):
-        self.end = time.time()
-        self.elapsed = self.end - self.start
 
 
 # From https://www.saltycrane.com/blog/2009/11/trying-out-retry-decorator-python/
